Some changes and fixes

+++ app/Http/Controllers/TaskController.php
@@ -34,6 +34,9 @@ class TaskController extends Controller
     public function edit(Request $request, $id)
     {
         $task = Task::findOrFail($id);
+        if ($task->user_id !== Auth::id()) {
+            return response()->json(['success' => false, 'message' => 'Forbidden'], 403);
+        }
         $task->update([
             'title' => $request->input('title'),
             'description' => $request->input('description'),
@@ -48,6 +51,9 @@ class TaskController extends Controller
     public function toggleChecked($id)
     {
         $task = Task::findOrFail($id);
+        if ($task->user_id !== Auth::id()) {
+            return response()->json(['success' => false, 'message' => 'Forbidden'], 403);
+        }
         $task->isChecked ^= true;
         $task->save();
         return response()->json([
@@ -59,6 +65,9 @@ class TaskController extends Controller
     public function toggleDelete($id)
     {
         $task = Task::withTrashed()->findOrFail($id);
+        if ($task->user_id !== Auth::id()) {
+            return response()->json(['success' => false, 'message' => 'Forbidden'], 403);
+        }
         if ($task->trashed()) {
             $task->restore();
         } else {
@@ -73,6 +82,9 @@ class TaskController extends Controller
     public function restore($id)
     {
         $task = Task::onlyTrashed()->findOrFail($id);
+        if ($task->user_id !== Auth::id()) {
+            return response()->json(['success' => false, 'message' => 'Forbidden'], 403);
+        }
         $task->restore();
         return response()->json([
             'success' => true,
@@ -83,6 +95,9 @@ class TaskController extends Controller
     public function delete($id)
     {
         $task = Task::findOrFail($id);
+        if ($task->user_id !== Auth::id()) {
+            return response()->json(['success' => false, 'message' => 'Forbidden'], 403);
+        }
         $task->delete();
         return response()->json([
             'success' => true,
